cpu-boost: Introduce Dynamic Stune Boost
Dynamic schedtune boosting occurs on-the-fly during interactions, such as touch input. This idea was conceived because of an existing battery drain issue whereby setting a permanent schedtune boost would cause boosting to occur even when the phone was idling. For example, if a user sets the /dev/stune/top-app/schedtune.boost to 10, the device may not idle at the lowest frequency step, which can lead to higher idle and active drain rates. Thus it should be better on the battery to boost only during interactions.

The tunable can be found in /sys/module/cpu_boost/parameters/dynamic_stune_boost.

Signed-off-by: joshuous <joshuous@gmail.com>

# Conflicts:
#	drivers/cpufreq/Kconfig
#	drivers/cpufreq/cpu-boost.c
#	include/linux/sched.h

# Conflicts:
#	include/linux/sched.h